evo2-sae: move probing harness into the evo2_sae.eval.probing package; drop CI lane
Relocate #1636's probe harness from scripts/ into the evo2_sae.eval.probing package (alongside the
#1629 primitives, now the package __init__):
  scripts/{labelers,evo2_buffer,annot_tracks,euk_windows,probe,probe_loss_recovered}.py
    -> src/evo2_sae/eval/probing/*.py
Fix imports to package-relative (from . import labelers; from .evo2_buffer import ...) and pull the
primitives from evo2_sae.eval.probing; loss_recovered stays in the shared sae lib. Re-point the tests
at the package (drop the sys.path-into-scripts/ hack). Remove the CPU CI lane (defer; run via .ci_build.sh
+ pytest). Reparented onto the moved #1629.
